move mainPages to pages
