+++ src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../layouts/Layout.astro";
-import WrapperPage from "../components/MainPages/wrapper";
+import WrapperPage from "./MainPages/wrapper";
 
 ---
 <Layout>
